+++ broadcastworker/main.go
@@ -5,11 +5,11 @@ import (
 	"time"
 
 	"github.com/Lavizord/checkers-server/config"
+	"github.com/Lavizord/checkers-server/logger"
 	"github.com/Lavizord/checkers-server/messages"
 	"github.com/Lavizord/checkers-server/redisdb"
 )
 
-var name = "BroadCastWorker"
 var redisClient *redisdb.RedisClient
 
 func init() {
@@ -17,25 +17,38 @@ func init() {
 	redisConData := config.Cfg.Redis
 	client, err := redisdb.NewRedisClient(redisConData.Addr, redisConData.User, redisConData.Password, redisConData.Tls)
 	if err != nil {
-		log.Fatalf("[BroadcastWorker][Redis] Error initializing Redis client: %v", err)
+		logger.Default.Fatalf("[BroadcastWorker][Redis] Error initializing Redis client: %v", err)
 	}
 	redisClient = client
+	logger.Default.Infof("broadcastworker initialized...")
 }
 
 func main() {
-	log.Printf("[BroadcastWorker] - Broadcasting room stats...\n")
+	logger.Default.Infof("broadcastworker starting up...")
 	ticker := time.NewTicker(time.Duration(config.Cfg.Services["broadcastworker"].Timer) * time.Second)
-
+	logger.Default.Infof("broadcastworker loaded timer configuration...")
 	defer func() {
 		ticker.Stop()
 		if redisClient != nil {
 			redisClient.CloseRedisClient()
 		}
 	}()
+
+	logger.Default.Infof("broadcastworker loaded timer configuration...")
 	for range ticker.C {
-		msg, _ := messages.GenerateGameInfoMessageBytes(redisClient)
+		keyCheckers := "game_info:{BatalhaDasDamas}"
+		keyChess := "game_info:{BatalhaDoChess}"
+		msg, _ := messages.GenerateGameInfoMessageBytes(redisClient, "BatalhaDasDamas")
 		// Publish the message
-		err := redisClient.Publish("game_info", msg)
+		err := redisClient.Publish(keyCheckers, msg)
+		if err != nil {
+			log.Printf("[BroadcastWorker] Error publishing message: %v\n", err)
+		} else {
+			//log.Printf("[BroadcastWorker] Published room aggregates")
+		}
+
+		msgChess, _ := messages.GenerateGameInfoMessageBytes(redisClient, "BatalhaDoChess")
+		err = redisClient.Publish(keyChess, msgChess)
 		if err != nil {
 			log.Printf("[BroadcastWorker] Error publishing message: %v\n", err)
 		} else {
